Fix RuntimeError/Issue #67
It seems this is caused by line 551 and the fact that in Python3 .items() creates an iterator object unlike Py2 which creates a copy of the object.  Using dict(ri_rel).items() creates an iterator of a copy of ri_rel allowing 551 to modify the original ri_rel without triggering an exception.

Tested on my own system without issues.